添加处理注解消息器，协议注册器，等重要工具类开发

+++ src/main/java/com/twjitm/common/logic/handler/AbstractBaseHandler.java
@@ -28,7 +28,7 @@ public abstract class AbstractBaseHandler implements BaseHandler {
 
     }
 
-    public Method getMethod(long commid) {
+    public Method getMethod(int commid) {
         return handlerMethods.get(commid);
     }
