+++ src/main/java/com/twjitm/common/logic/handler/BaseHandler.java
@@ -3,5 +3,5 @@ package com.twjitm.common.logic.handler;
 import java.lang.reflect.Method;
 
 public interface BaseHandler {
-    public Method getMethod(long commid);
+    public Method getMethod(int commid);
 }
